Fix so that generated sims don't need to be modified to support `repeat`
Prior to this commit, the pattern that I was using to allow sims
generated by the Gatling Recorder to support programmatic/config-driven
`repeat` functionality was to edit each one by hand and add a
constructor argument for the desired number of repetitions.

Thanks to some help from the gatling mailing list, I was able to
cobble together a hack in `ConfigDrivenSimulation` that removes
the requirement for that particular hand-edit for each generated
simulation.

This commit also removes those hand-edits from the two existing
sims.