Fix the class assignment code in the master config script
Prior to this commit, there were a few bugs in the code
that was assigning classes to a node.  First, the assignment
was happening in a loop over the desired classes... so if there
were no desired classes, then no assignment would occur at all.

Second, we were using the `classes` rake task inside the loop
to assign a single class.  Since that task overwrites the
entire list of classes for a node, the end result would be
that the node would have exactly one class assigned to it, and it
would be whichever class appeared last in the list of desired
classes.