Don't sleep 30 minutes after final repetition in long-running
Prior to this commit, we were sleeping for 30 minutes after
every repetition of the agent requests in the long-running
simulation.  This sleep is not necessary after the final
repetition, and was adding 1-3 hours of unnecessary run time
to the jenkins jobs.  Getting rid of that last unnecessary
sleep should allow us to run through the jenkins jobs
much more quickly.