Use more descriptive names for requests in node sims
Gatling auto-generates label strings to identify the various
requests that make up a simulation.  They are generic and
auto-incrementing ("request_1", "request_2", etc.).  These
show up in the final report HTML, and make it a little
more difficult than it needs to be to distinguish between
the different kinds of requests and understand what's going
on.

This commit simply gives the requests more descriptive names
so that they way they show up in the UI will be more meaningful.