(feature) Add Collect Artifact Setup Step
Prior to this commit, we needed a way to keep track of some important
operating system data while running our gatling puppet load tests. This
commit adds a ruby script to gather that data from facter run on the
master, which is then parsed out for specific data points we are
interested in, then both the full facter output and the important data
is written to a text file and csv file respectively.