(feature) Add json parse of gatling config
This commit adds the functionality to get all of the configuration set
in Jenkins, with the required numbers of # node instances and # of
repetitions.

Prior to this commit, our setup files needed a way to share the gatling
config parsed from json files. This commit adds a helper to accomplish
that.