Add `set -e` to cobbler script
I've noticed that on some of our cobbler runs, the ssh connection
to the cobbler controller box (`ull.delivery.puppetlabs.net`) fails,
which means that the command to tell the cobbler slave box to
reinstall its OS is never executed.  When this happens, it doesn't
cause the job to fail, and we just keep on moving.

This commit adds a `set -e` to the cobbler script in hopes of
forcing a failure when the controller box cannot be reached.