Allow for a target machine to be specified in the json file
Prior to this commit we were unable to specificy what hardware target
the a job would use. It was actually hard-coded in the
cobbler_provision.sh script.

Now there is a new field "target" in the json file per job that is used
as the Cobbler Host Profile Target to deploy upon. For this initial
commit, I have set all of them to the original hard-coded value found in
the provisioning script, but now we could potentially use other targets
since at times we have mulitple hardware options available.

This will also raise an error in bin/perf if a target is not found in
the json job file.