Set hostname to dns name of perf02
When this code was initially created, the perf nodes did not have
dns entries.  We set the hostname to 'puppet' and simply relied on
using the IP directly.  We require the actual hostname for the new
PE3.2 installer.  Next steps would be to provide for a method for
setting IP and hostname dynamically and depricate the use of
'target' in favor setting and using the hostname consistantly.