(SERVER-780) Validate node classification
This commit adds a beaker script that will query the master's HTTP API
and validate that each node will receive the correct classes during the
catalog request.

This beaker script requires that each node already has an SSL
certificate and private key on the master. Unfortunately there is no
automation for putting the certificate and key on the master yet,
see SERVER-787.

This validation script will likely be executed as the very last thing
prior to launching Gatling.