(SERVER-854) Dev/ improvements; start of OPS job
This giant commit contains many fixes to the existing dev/ scripts so
they actually work properly again (they're still very specific to CentOS
6 though).

This also contains various adjustments to the new automation steps we
have in jenkins-integration, as this is the first time we're actually
trying to use them. Some reconciliation between new and old was
necessary, and there are certinaly more improvements that can be made.

Lastly, this commit contains the beginnings of the OPS deployment JJB
job. The JJB job uses the new jenkins-integration automation steps and
defines it's own specific step to clone the new
test-catalogs-large-files repository and untar the OPS environments into
code-staging.

The OPS job is not yet complete, and will likely need updates to the
ops-scenario.json and ops-node.json files as we continue development.