Move feeder logic into ConfigDrivenSimulation
This commit moves the feeder logic into the ConfigDrivenSimulation
class.  This means that the CSVs for feeding node names are no
longer necessary.

However, when this commit is merged, any existing scenarios that
use feeders will probably be broken.  Also, any of the automation
logic that expects to use the certnames from the node config json
files to do classification will not work properly (though this
was probably already the case for any of the feeder simulations).

Probably need to:

* Fix existing feeder sims to work with this approach
* Modify the NodeFeeder to support having the node prefix string
  be read from the node config json, instead of using the hard
  coded value 'node' as we are currently doing.  Without that change,
  any multi-node scenarios will have collisions in their certnames.