(MAINT) take 2, raw recordings for burnside large files
We discovered that JGit has a memory optimization that kicks
in for files over 50mb.  Redoing these recordings to use
files that are over that threshold.  Also pursuing some related
tasks in the PE code that will allow us to configure this threshold.

This commit just has the raw recordings for the sake of git history.
They need to be modified to work with g-p-l-t.