(MAINT) Add a catzero simulation for requests using msgpack
This commit adds a catalog zero simulation for PE Couch which utilizes
msgpack support, including msgpack in client Accept headers and a report
payload encoded with msgpack.

In order for this to be playable, the following changes also needed to
be made to the target master:

1) Enable msgpack on both the agent and server - see
   https://docs.puppet.com/puppet/latest/reference/experiments_msgpack.html.

2) Make report submissions unique (uuid and timestamp) in Ruby Puppet

It might be possible to do that Gatling-side by parameterizing the
report payload but this PR doesn't include any changes which take care
of that automatically.  Instead, it relies upon changes being made in
the Ruby Puppet code on the target master.

For example, this could be added to do_save() in
lib/puppet/network/http/api/indirected_routes.rb:

   require 'securerandom'

   if indirection.name == :report
     sent_object.host = key
     sent_object.transaction_uuid = SecureRandom.uuid
     sent_object.instance_variable_set(:@time, Time.now)
   end