(SERVER-1464) Use clientcert instead of trusted.certname for classification
Prior to this commit, the code that was doing the NC classification was using
the `trusted.certname` field to assign nodes to the node group.  However,
when running gatling sims, the simulated agents will probably all be sharing
the same cert, and it probably won't match the desired pattern.  The `clientcert`
field is a better fit for our current setup.