(SERVER-1452) Copy hiera config from source to target directory on SUT
This commit enables a job configuration to include a source location,
`hiera_config_source_file`, from which a hiera.yaml file will be copied
to the target location where Puppet expects it to be.  The datadir to be
used in the hiera.yaml file is also now configurable on a per-job basis
via the `hiera_config_datadir` key.