(SERVER-1471) Move scenario config out of job params
This commit moves the logic for specifying the
PUPPET_GATLING_SIMULATION_CONFIG setting from a Jenkins workflow job
parameter down into a job parameter at the level where the pipeline is
being constructed.  This was done since a multipass pipeline needs to
allow for the simulation config to be unique per "job" it runs.