(SERVER-1563) Leave pe_dir option off the hostgen command line if blank
Previously, if the pe_dir option passed down to the 010_setup_beaker.sh
was empty, it was being passed along to beaker-hostgenerator as
"--pe_dir ''".  This caused a value of '' to be used in the generated
host file, leading a PE install later on to fail because of the bogus
directory location.

This commit removes the pe_dir setting from the beaker-hostgenerator
commaand line when it is blank.  This allows hostgenerator to insert its
own default for the pe_dir field, which "should" be appropriate for a
released (non-latest internal-only) PE build.