(SERVER-1662) Fix cron collisions, introduce server_config
This commit introduces a new config file, `server_config.json`.
In it we can list out hostnames of jenkins servers, and set
attributes for them.  The attributes will be made available to the
`JobDSL.groovy` scripts via a local variable called `serverConfig`.

This allows JobDSL configurations to do things like conditionally
modify a job based on whether it is being created in a dev environment
vs. a production environment.