(SERVER-1790) Skip non-existent cent7 puppet-agent builds
Previously, the logic which tried to determine which 'latest'
puppet-agent release build to install would assume that a cent7 build
was available on the repo server if its parent directory existed.  This
will sometimes not be true - e.g., where some artifacts may be built up
over the course of a few days near a release.

The logic for determining the latest puppetserver release skips over any
build where the repo folder exists but the cent7 build itself does not.
This commit reworks the logic for determining the latest build to be
more common for both puppet-agent and puppetserver.  As a result,
non-existent puppet-agent builds will be skipped over like puppetserver
ones have been.