(SERVER-1790) Enable tk-auth for latest/master OSS Puppet Server runs
Previously, all OSS Puppet Server jobs assumed that the
use-legacy-auth-conf setting was set to 'true' and so only tried to
reconfigure the Ruby Puppet auth.conf file to enable the permissive
access needed to perform multi-node runs.  As of Puppet 5, though,
use-legacy-auth-conf will be set to 'false' by default.  This commit
updates the server config to assume that for any builds being tested
from the Puppet Server master branch, the tk-auth configuration file is
setup for permissive access.