(SERVER-1875) Add oss puppetserver 2.x and 5.0.x gatling jobs
This commit adds oss puppetserver jobs for the 2.x and 5.0.x branches.
In order to allow for the jobs to specify a version pattern which
matches the artifacts that each branch builds, this commit also
introduces support for interpreting the content of the version string as
a regular expression whenever it contains the text "SNAPSHOT" in it.  In
this case, the latest build matching the regular expression should be
installed.