(SERVER-2034) Update puppet-agent jenkins link
At some point the jenkins job that was being used changed names and
the method that calls out to get the latest passing agent SHA stopped
working. This commit updates the method to use a more stable URL that
also requires less navigation to get to the desired hash element.