(QA-3410) Stop forcing filesync
The force of filesync in the file that is being removed was actually deleting all of the code that r10k had deployed.  File sync was needed for cases where we have compile masters.  But for that we need to move up to using code manager since that is what the customer do.  For Foss we will have to come up with a solution as there is no code manager...