(SERVER-2071) Don't care if background script was already stopped
Earlier today we saw a frivolous failed status for a job because this
task attempted to stop a background script, but that PID did not exist,
for reasons we couldn't determine. Since the harm in leaving that
running is minimal (nodes get wiped before reuse in most cases) but
results don't show up properly if the job is marked as a failure, we
want to ignore the case in which the PID did not exist and just
continue.