(BKR-487) host_preserved.yml fix
changed gemfile to pick up a special version of beaker that has a fix to convert the AWS objects to hashs.
This will enable the host_preserved.yml to be read in, but will not enable it to use anything that leveraged
the aws instance object... like killing the instances.

This is a temp fxi until QE does the longer term fix and makes a release.