(SERVER-2078) Add HEAP_OVERRIDES, JAVA_ARGS_ADDITIONS param to jobs
Previously changing gatling job parameters such as heap settings or
which garbage collector to use was impossible. This commit adds two new
parameters to all of the gatling jenkins jobs that will allow us to
override heap settings and to supply additional java args to jobs such
as garbage collector.