(maint) Add a test job
When testing out changes to our lovely jenkins jobs, it is nice to not
need to wait two or more hours to find out that groovy thinks your
double quoted string should be a dsl method. This commit gets far faster
feedback in groovy shenanigans by adding a test job that runs a single
iteration with a 10 second ramp up and sleep period.