(maint) Set a heap size of 2gb unless specified
We've run into a couple of cases where the heap size was not specified
and the jvm defaulted to 1/4 of system memory. To avoid that, this
commit makes explicit a default of 2gb heap for all tests that don't
specify it. HEAP_OVERRIDES will still win, so this can be altered at the
job build screen.