(SERVER-2091) Always be gc logging
So far it has been advantageous to collect GC logs from our gatling runs
and it has a negligible perf impact, so this commit updates all of the
jobs to archive the gc log if it exists and also adds gc logging java
args as the default java args for all jobs. Archive collection is now
configured to skip files that don't exist instead of failing.