(SERVER-2108) Add dynamic gatling job
Currently in order to test new combinations of node count, duration, or
catalog size a new job and scenario must be constructed. This is slow
and tedious, so this commit adds a new jenkins job that constructs a
custom gatling scenario on the fly and uses it to kick off a job against
a node config for the desired catalog size.