(SERVER-2108) Fix scenario file path
Previously the custom scenario file was set to just the filename, and
not the full path to the scenario file, when being generated. This
unfortunately meant that none of the functions could find the file
later. This commit prepends the path to the filename to ensure it can be
found.