(SERVER-2017) Razor-ify smaller blades
Adding tags for mac addresses of current
perf-bl{11,12,13}.delivery.puppetlabs.net. This also means they believe
their hostname is perf-b1{3994,3995,3996}, respectively, because Razor
identifiers are the only thing we can use there, we can't specify
particular IDs for nodes, Razor determines those. The 3 nodes were
registered with Razor manually (the acceptance key will get you in) and
should Just Work now. The "small" "sut_size" metadata was added so that
we could have different disk partitioning behavior on the larger suts
that these (the original ones have two disks and more elaborate
partitioning; with these we just use autopart); this is done in the
kickstart file in the razor task in puppetlabs-modules.

From some experimentation I've found that for the current broker
(specified in puppetlabs modules) will allow you ssh cert access with
the gatling key (available on int-resources, see PR #130), but not with
the acceptance key.

Note for posterity: mac addresses in Razor tags are case-sensitive,
apparently.