(SERVER-2017) Job to set up new server
Create a new kind of pipeline that can be used to spin up a puppetserver
instance for some hands-on testing. This is basically a gatling run
without gatling or cleanup (so any desired artifacts will have to be
archived manually).

Still want to create a tool to simulate agent load without gatling, but
not sure that kicking it off immediately from this job is the right way
to go.