(SERVER-2108) Update jar customization to handle no param case
Because the JRUBY_VERSION parameter was only added to the
puppetserver-infinite job and because we are on an older version of the
pipelines plugin, jobs that don't define the JRUBY_VERSION parameter
fail with a MissingPropertyException. In our current version of the
pipelines plugin, the recommended approach seems to be a try/catch
block. When we update to a more recent version this hackery can be
unwound.