(SERVER-2123) Update job to run gatling against puppet branch
In order to be able to better test how some changes to puppet affect
performance, this commit updates the configurable job so that it can run
gatling against a puppet branch or ref. It clones and checks out the ref
locally and rsyncs it on top of the puppet-agent installation, which is
gross, but ultimately far simpler and faster than crafting a vanagon
pipeline.