(maint) Use `hostname` to designate puppet's server
The value of beaker's `master` variable is not always the same as a perf
node's `hostname`. When invoking puppet on these hosts, we need
to use the hostname as the server to contact, and not the external alias
given to beaker. This commit updates the `puppet agent` invocation to use
`hostname`, which works in both cases.