(QA-3456) Use Gatling feeders to simulate facts change
With this change, facts are no longer a hard-coded string inside the .scala
file itself. Most of the hard-coded facts string has been moved to external
file (IrvingFactsChange_facts_raw.txt) with a small "header" inline in order
to position our custom facts so they are in the correct part of the URL-encoded
JSON sent to Puppet Server.

There are two classes of changeable facts. The first set
(IrvingFactsChange_custom_facts.csv) is for facts that change with every agent
run. There should be enough lines in the file to supply data for every agent
run. In other words, for 600 agents running four hours (8 runs per agent) there
should be 4800 + 1 lines. If there aren't enough lines for the full Gatling job
run, the test will produce an error. This is a feature, not a bug.

The second class of facts are unique to each agent, but do not change during
the run  (user-files/facts/IrvingFactsChange_custom_hostf.csv). There should be
one line for every agent configured for the test.

The facts are all pre-generated before the test in order to remove the
possibility of slowing down the test driver side. The facts are all loaded into
memory, so facts on the order of gigabytes will probably produce memory errors
on the driver node. Facts should be in the following format:

%22customfact1%22%3A%22KO8mxN5FntvC6NwW%22%2C

To disambiguate the above:

%22 customfact1 %22 %3A %22 KO8mxN5FntvC6NwW %22 %2C
 "  customfact1  "   :    " KO8mxN5FntvC6NwW  "   ,

The key "customfact1" as well as its value "KO8mxN5FntvC6NwW" do not really
matter, as long as the values in bulk contain enough data to be effectively
random.

A simple script(s) can be utilized to generate facts.

echo "Fact" >> tmp.csv for x in `seq 1 4800` ; do echo "\
%22customfact1%22%3A%22`pwgen -s -1 16`%22%2C\
%22customfact2%22%3A%22`pwgen -s -1 24`%22%2C\
%22customfact3%22%3A%22`pwgen -s -1 32`%22%2C\
" >> tmp.csv ; done

echo "HostFact" >> tmp.csv for x in `seq 1 600` ; do echo "\
%22hostfact1%22%3A%22`pwgen -s -1 16`%22%2C\
%22hostfact2%22%3A%22`pwgen -s -1 24`%22%2C\
%22hostfact3%22%3A%22`pwgen -s -1 32`%22%2C\
" >> tmp.csv ; done