(SLV-235) Fix recordings for no more MCO
removed MCO requests for PerfTestLarge recording.
updated puppet and ruby versions in the recording
enables a rake task to run a presuite that uses perf_helper to setup the metrics node so it can be used to record an agent run.
modified launch gatling to stop requiring it be a git repo and to remove code that was moved to the pre-suite
removed request count assertion from ApplestoApples as it is totally dependent on the recording.