(SLV-266) Fixed baseline comparison
converted the big_query key that was a symbol to a string for use in the process_results hash.
Also added a line so that sims that have baseline asserts don't fail when baseline_ver isn't set.
updated acceptance request count to 60 to account for no more mco
removed response time assertion since the test is very short it will vary a lot.