(SLV-318) Remove credentials from scooter call
This commit removes the user/password credentials when creating a
`ConsoleDispatcher` object.  With the release of scooter 4.3.1, a ca
cert acquisition bug was fixed that causes the auth to fail if it is also
passed credentials.  These credentials have been removed in order for
the dispatcher will use its native cert-based auth.