Get puppetserver service from BEAKER_INSTALL_TYPE
This commit removes the environment variable usage of
`PUPPET_SERVER_SERVICE_NAME` and updates the
`get_puppet_server_service_name_from_env` method to derive the name from
the `BEAKER_INSTALL_TYPE`.  The value of `pe-puppetserver` is returned
if the install type is set to `pe` otherwise the foss value of
`puppetserver` is returned.