(SLV-329) Fix baseline variance math
This commit removes the conditional for calculating the variance
of 'cpu' process values.  These values are never evaluated due to
`next` condition prior to the conditional used to evaluate them.

This commit also updates the conditional for non-cpu values to
correct the hard-coded 10% value to use the variable
`max_baseline_variance` to determine the calculation.  This variable
is set to 15% for orch_mem and otherwise set to 10%.