(SLV-414) Add scale test script
This commit adds a bash script for automating running scale testing
against the defined reference architectures defined in the PE docs

https://puppet.com/docs/pe/latest/hardware_requirements.html

This script will run a single iteration of the "10 or fewer" architecture.  It
will then run the following iterations on three different EC2 instance
definitions to test the scale of the "Up to 4,000" architecture options.

1. Cold start run provisioning the EC2 instances.
1. 4 more cold start iterations on the instance created above.
1. 5 warm start iterations on the instances created above.

The "Monolithic with compilers" architecture is not executed.