(SLV363)Task take public github repo as source
updated the gplt::create_control_repo_from_production_env task to be able to accept a public github repo to be the source of the local repo it creates.  Also accepts a branch, which it will convert to the production branch in the new repo.
This allows for easy creation of a repo to use with the simulation from any branch of any public control repo.
The intent is to then point code manager at this repo as part of setup.

updated docs on the task as well