use yum for scala install to make it resilient
This commit updates the perf helper step for installing scala to use
`yum` rather then `rpm -i`.  Prior to this commit, running the setup
step would fail if the scala rpm had been previously installed.  Using
yum allows the installation to be gracefully detected without failure if
an update is not required.