(SLV-430) Added metrics module and script
created the metrics module, and a script for future commit's to call from a plan.
Added spec test frame work, but no real tests.
The script finds all the files in the metrics dir that are relevant to a time window and tars them up into a
single tar file.
This is part 1.
Part 2 will be spec tests.
Part 3 will be the bolt plan that calls the script and bring back the resulting file.

Also, told rubocop to stop complaining about explicit return statements.