(SLV-430) Move CMF out of bolt, add one spec test
turned out that writing a plan to run the script and get the file is not possible with bolt as is.  So pivoting to make it a normal util script
This just moves it and it files, as well as adjusts spec setup to work in the new place.

It also adds a spec test for json and tar files that don't match the expected pattern... the code should ignore them...
added a change to the code to do so.