(SLV-533) Use params if nodes exist
This commit replaces the string representation of the pe_xl params with
a hash.  The hash is compacted to only contain keys with values.  So, if
any of the defined node role types have not been provisioned, then the
corresponding parameter will not be created.

The hash is then transformed into a pretty-printed JSON string using the
JSON library to be written to the output_dir.