(SLV-355) Use CMF to get tarball of metrics
Added code to generate a start_epoch and end_epoch file with the epoch of the, you guessed it, start and end of the test.
Added code to use the epochs in those files as arguments to the collect_metrics_files script to generate a tarball of metrics, and scp it into the results dir.