(SLV-573) supress new lines in csv file
split_atop_csv_results splits the file into 2 files.  1 for the top section and one for the bottom.
But it was sending the empty line to the first file, and then csv lint failed on that blank line.
So I added code to supress the printing of blank lines in split_atop_csv_results