(SLV-574) Add privatebindir to master if not avail
This commit updates the `get_perf_result` method in the
`perf_run_helper.rb` file.  It looks up the bindir of the installed
puppet on the master if the `privatebindir` value has not been set.

Prior to this commit, the call to `ruby /root/collect_metrics_files.rb`
on the master would fail on a pe_xl deployed environment because beaker
was not used to install puppet.  Without installing puppet, beaker does
not assign a value to `privatebindir` on the host.  This results in ruby
not being available on the PATH, so calling ruby fails.