(SLV-647) Ensure all puppet changes post inst
This commit adds a `run_agent_until_no_change` helper to
`setup/helpers/perf_helper.rb` and updates the `30_final_puppet_run`
pre-suite to use it to ensure that the puppet agent is run repeatedly
until all changes are applied and the exit code is `0`.

The method replaces previous code in `10_puppet_infrastructure_tune`.

Spec tests for this helper are added.