(SLV-520) Add gplt CLI and validate2baseline
This commit adds the gplt CLI command and the `validate2baseline` to
provide the ability to validate performance results data against
baseline data.

The subcommand requires the following options be provided:
* baseline - Baseline data to compare to (e.g. 2019.1.0)
* results_dir - Path to gatling results to extract data from

The following option is optional and defaults to "apples to apples".
This is required for the big query lookup of the desired baseline data.
* test_type - Test type used to create baseline data

This interface is a scaffolding only.  The required logic for the
`validate_results_to_baseline` method is not implemented yet.