(SLV-691) Create system log archive in results dir
This commit copies the system logs from the SUT into a tar archive in
the results directory on the runner.

Prior to this commit, the archive was created at the root of the
runner's gplt checkout.  With this commit, the archive is embedded in
the results directory for the executed perf/scale run.

The archive is no longer stored separately from the results directory.
As a side effect, Jenkins will no longer receive a separate archive for
the system logs.  Instead, they can be found in the archive of the
results themselves which is downloaded at the conclusion of the Jenkins
run.