(SLV-520) Archive logdir on SUT and copy down
This commit changes the implementation of the `copy_system_logs` helper.
It now archives the puppet logdir on the SUT and then copies the archive
from the SUT to the runner rather than copying the logdir tree and
archiving on the runner afterward.

This change is being made to reduce the amount of data to be copied from
SUT to runner.