(SLV-702) handle duplicate params in conf files
Conf files can have the same parameter defined multiple times.  If they are the last one wins.  The code change here sets us up to get the last one instead of the current one.
That change made it also a good time to remove the exclusion mechanism since it would have to have been refactored to continue to function.
Also Fixed the reporting on the max jruby instances to be NA if there was no puppetserver conf file instead of reporting the calculated default jrubies.