(SLV-656) Add save_average_compile_time helper
This commit adds the `save_average_compile_time` helper to to the
PerfRunHelper module.  This helper queries puppetdb for the
"config_retrieval" and "total" values associated with metrics data from
the reports endpoint.  This query is constrained to a time range bound
by the start_epoch and end_epoch instance values associated with the
code at runtime.

Spec tests for this helper are added as well.