(SLV-708) Skip improvement variances
This commit adds a filter in the `validate_baseline_data` helper to skip
failing variances where the variance is an improvement from the
baseline.  These have not been used as a failing condition in CI in the
past, so we will skip them for the moment.  This commit should be
reverted when a process for addressing such variances has been defined.