(SLV-656) Rename save_average_compile_time helper
This commit renames the `save_average_compile_time` helper to
`save_average_transaction_time`.  This change reflects the nature of the
data collected which is not directly associated with server side compile
times, but the agent submitted reports for transaction times via
catalog request.

The payload saved to the file is updated to be a JSON data structure
rather than a single number.