(SLV-609) Download pe tarballs for pe_xl
This commit adds logic to the `pe_xl:provision` rake task to download the
package for the given BEAKER_PE_VER.  The pe_xl plans will use a local
`tar.gz` file for the given version from the `/tmp` directory if it is
available.  This change determines whether the given version is a
release or development version, downloads the tar file, and gzips it if
necessary in order to provide it to pe_xl.