(SLV-610) Separate compiler csv data
This commit updates the file creation logic in the
`extract_puppetserver_metrics` helper to create discrete files per
puppetserver host.

In the scenario of using load balanced compile masters for gatling node
catalog compilation the primary master puppetserver is not utilized for
catalog compilation.  This results in empty csv data.  For this reason,
the generation of the puppetserver metrics extraction files has been
updated to no longer throw an exception in the event of invalid csv
data.  A wrapper method called `valid_csv?` has been added to get a
boolean result as to the validity of a given csv file.  In the case of
an invalid file the extraction process skips over further processing of
the file and moves on to the next step/iteration of the extraction
process.